[#3081] refactor: move `AxonServerRegistration` to correct package
