+++ axon-server-connector/src/main/java/org/axonframework/axonserver/connector/AxonServerRegistration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.axonserver.connector.command;
+package org.axonframework.axonserver.connector;
 
 import org.axonframework.common.Registration;
